go/consensus/tendermint/apps/roothash: Move transactions to its own file
