+++ go/consensus/tendermint/apps/roothash/roothash.go
@@ -4,7 +4,6 @@ package roothash
 import (
 	"bytes"
 	"encoding/binary"
-	"fmt"
 
 	"github.com/pkg/errors"
 	"github.com/tendermint/tendermint/abci/types"
@@ -478,134 +477,6 @@ func (app *rootHashApplication) FireTimer(ctx *abci.Context, timer *abci.Timer) 
 	}
 	for _, pool := range rtState.Round.ComputePool.GetTimeoutCommittees(ctx.Now()) {
 		app.tryFinalizeCompute(ctx, runtime, rtState, pool, true)
-	}
-
-	return nil
-}
-
-type roothashSignatureVerifier struct {
-	runtimeID common.Namespace
-	scheduler *schedulerState.MutableState
-}
-
-// VerifyCommitteeSignatures verifies that the given signatures come from
-// the current committee members of the given kind.
-//
-// Implements commitment.SignatureVerifier.
-func (sv *roothashSignatureVerifier) VerifyCommitteeSignatures(kind scheduler.CommitteeKind, sigs []signature.Signature) error {
-	if len(sigs) == 0 {
-		return nil
-	}
-
-	committee, err := sv.scheduler.Committee(kind, sv.runtimeID)
-	if err != nil {
-		return err
-	}
-	if committee == nil {
-		return errors.New("roothash: no committee with which to verify signatures")
-	}
-
-	// TODO: Consider caching this set?
-	pks := make(map[signature.PublicKey]bool)
-	for _, m := range committee.Members {
-		pks[m.PublicKey] = true
-	}
-
-	for _, sig := range sigs {
-		if !pks[sig.PublicKey] {
-			return errors.New("roothash: signature is not from a valid committee member")
-		}
-	}
-	return nil
-}
-
-func (app *rootHashApplication) commit(
-	ctx *abci.Context,
-	state *roothashState.MutableState,
-	id common.Namespace,
-	msg interface{},
-) error {
-	logger := app.logger.With("is_check_only", ctx.IsCheckOnly())
-
-	rtState, err := state.RuntimeState(id)
-	if err != nil {
-		return errors.Wrap(err, "roothash: failed to fetch runtime state")
-	}
-	if rtState == nil {
-		return errNoSuchRuntime
-	}
-	runtime := rtState.Runtime
-
-	if rtState.Round == nil {
-		logger.Error("commit recevied when no round in progress")
-		return errNoRound
-	}
-
-	latestBlock := rtState.CurrentBlock
-	blockNr := latestBlock.Header.Round
-
-	defer state.SetRuntimeState(rtState)
-
-	// If the round was finalized, transition.
-	if rtState.Round.CurrentBlock.Header.Round != latestBlock.Header.Round {
-		logger.Debug("round was finalized, transitioning round",
-			"round", blockNr,
-		)
-
-		rtState.Round.Transition(latestBlock)
-	}
-
-	// Create storage signature verifier.
-	sv := &roothashSignatureVerifier{
-		runtimeID: id,
-		scheduler: schedulerState.NewMutableState(ctx.State()),
-	}
-
-	// Add the commitments.
-	switch c := msg.(type) {
-	case *roothash.MergeCommit:
-		for _, commit := range c.Commits {
-			if err = rtState.Round.AddMergeCommitment(&commit, sv); err != nil {
-				logger.Error("failed to add merge commitment to round",
-					"err", err,
-					"round", blockNr,
-				)
-				return err
-			}
-		}
-
-		// Try to finalize round.
-		if !ctx.IsCheckOnly() {
-			if err = app.tryFinalizeBlock(ctx, runtime, rtState, false); err != nil {
-				logger.Error("failed to finalize block",
-					"err", err,
-				)
-				return err
-			}
-		}
-	case *roothash.ComputeCommit:
-		pools := make(map[*commitment.Pool]bool)
-		for _, commit := range c.Commits {
-			var pool *commitment.Pool
-			if pool, err = rtState.Round.AddComputeCommitment(&commit, sv); err != nil {
-				logger.Error("failed to add compute commitment to round",
-					"err", err,
-					"round", blockNr,
-				)
-				return err
-			}
-
-			pools[pool] = true
-		}
-
-		// Try to finalize compute rounds.
-		if !ctx.IsCheckOnly() {
-			for pool := range pools {
-				app.tryFinalizeCompute(ctx, runtime, rtState, pool, false)
-			}
-		}
-	default:
-		panic(fmt.Errorf("roothash: invalid type passed to commit(): %T", c))
 	}
 
 	return nil
